Add files via upload
needs 4th argument, puts all text files in folder "/alltexts/etcetc.txt"

+++ get_apis.py
@@ -23,13 +23,17 @@ except IndexError:
 datasetPath = sys.argv[1]
 
 try:
-	sys.argv[2]
+	sys.argv[3]
 except IndexError:
-	print("Need at least one file to get Sources/Sinks from")
+	print("Need both files to get Sources/Sinks from")
 	sys.exit(1)
-
+try :
+	sys.argv[4]
+except IndexError:
+	print("Enter the line number to start analysis. Enter 0 for all lines")
+	sys.exit(1)
 sosiClassDict = {}
-for i in range(2, len(sys.argv)):
+for i in [2,3]:
     input_file = sys.argv[i]
     with open(input_file) as f:
         for line in f:
@@ -49,46 +53,52 @@ for i in range(2, len(sys.argv)):
 
 if os.path.exists(os.path.join(maraPath, "data")) and os.path.isdir(os.path.join(maraPath, "data")):
 	shutil.rmtree(os.path.join(maraPath, "data"))
+if not os.path.exists("alltexts") and not os.path.isdir("alltexts"):
+	os.mkdir("alltexts")
 linenumber = 0;
+print(sys.argv[4])
 with open(datasetPath,"r") as apkpaths:
 	for pathe in apkpaths:
 		linenumber  = linenumber + 1
-		print("Line number:"+ str(linenumber))
-		apk = os.path.split(pathe)[1].strip("\n")
-		args = [os.path.join("./", maraPath, "mara.sh"),"-s",pathe]
-		subprocess.call(args, cwd=maraPath)
-		apkSmaliPath = os.path.join(maraPath, "data", apk, "smali/apktool/")
-		crt = open(apk+".txt","w")
-		for iroot, idirs, ifiles in os.walk(apkSmaliPath):
-			for name in ifiles:
-				with open(os.path.join(iroot,name),"r") as f:
-					for x in f:
-						if re.search('invoke-.*(Landroid).*',x):
-							print((str(re.findall('\w+',x.split('}')[1].split('-')[0])).translate(None, '[],\'')).replace(" ",".")[1:], end = " ", file = crt)
-							print(':', end = " ", file = crt)
-							print(re.findall('\w+',x.split('}')[1].split('-')[1])[0],file = crt)
+		if (sys.argv[4] != 0 and linenumber < int(sys.argv[4])):
+			print("skipping "+ str(linenumber))
+		else:
+			print("Line number:"+ str(linenumber))
+			apk = os.path.split(pathe)[1].strip("\n")
+			args = [os.path.join("./", maraPath, "mara.sh"),"-s",pathe]
+			subprocess.call(args, cwd=maraPath)
+			apkSmaliPath = os.path.join(maraPath, "data", apk, "smali/apktool/")
+			crt = open(apk+".txt","w")
+			for iroot, idirs, ifiles in os.walk(apkSmaliPath):
+				for name in ifiles:
+					with open(os.path.join(iroot,name),"r") as f:
+						for x in f:
+							if re.search('invoke-.*(Landroid).*',x):
+								print((str(re.findall('\w+',x.split('}')[1].split('-')[0])).translate(None, '[],\'')).replace(" ",".")[1:], end = " ", file = crt)
+								print(':', end = " ", file = crt)
+								print(re.findall('\w+',x.split('}')[1].split('-')[1])[0],file = crt)
 		#subprocess.call(["sort",apkname+".txt", "-o" ,"sorted"+apkname+".txt"])
 		#shutil.rmtree(os.path.join(maraPath, "data", apk))
-		crt.close()
-		final = open(os.path.join(os.path.split(pathe)[0] ,"final"+apk+".txt"),"w")
-		lineset = set()
-		with open(apk+".txt", "r") as cur1:
-			for x in cur1:
-				if(len(x.split(":")) > 1):
-					cname = str(re.findall("\w+",x.split(":")[0])).translate(None,'[],\'').replace(" ",".")
-					mname = str(re.findall("\w+",x.split(":")[1])).translate(None,'[],\'')
-					if(cname in sosiClassDict):
-						if mname in sosiClassDict[cname][0]:
-                                                    finwrite = ""
-                                                    for argsString in sosiClassDict[cname][1][mname]:
-							finwrite = "<" + cname + ": RETURN_TYPE " + mname + "(" + argsString + ")> (CATEGORY)\n"
-						        if finwrite not in lineset:
-							    final.write(finwrite)
-							    lineset.add(finwrite)
-		os.remove(apk+".txt")
-		final.close()
-		curapkdatapath = "data/"+apk
-		if os.path.exists(os.path.join(maraPath, curapkdatapath)) and os.path.isdir(os.path.join(maraPath, curapkdatapath)):
-			shutil.rmtree(os.path.join(maraPath, curapkdatapath))
+			crt.close()
+			final = open(os.path.join("alltexts","final"+apk+".txt"),"w")
+			lineset = set()
+			with open(apk+".txt", "r") as cur1:
+				for x in cur1:
+					if(len(x.split(":")) > 1):
+						cname = str(re.findall("\w+",x.split(":")[0])).translate(None,'[],\'').replace(" ",".")
+						mname = str(re.findall("\w+",x.split(":")[1])).translate(None,'[],\'')
+						if(cname in sosiClassDict):
+							if mname in sosiClassDict[cname][0]:
+								finwrite = ""
+								for argsString in sosiClassDict[cname][1][mname]:
+									finwrite = "<" + cname + ": RETURN_TYPE " + mname + "(" + argsString + ")> (CATEGORY)\n"
+									if finwrite not in lineset:
+										final.write(finwrite)
+										lineset.add(finwrite)
+			os.remove(apk+".txt")
+			final.close()
+			curapkdatapath = "data/"+apk
+			if os.path.exists(os.path.join(maraPath, curapkdatapath)) and os.path.isdir(os.path.join(maraPath, curapkdatapath)):
+				shutil.rmtree(os.path.join(maraPath, curapkdatapath))
